requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
tweepy>=4.14.0
Pillow>=10.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
newsapi-python>=0.2.7
